@@ -1,29 +1,32 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="coverlet.msbuild" Version="3.0.3">
+        <PackageReference Include="coverlet.msbuild" Version="3.1.2">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="DotNetEnv" Version="2.1.1" />
-        <PackageReference Include="FluentAssertions" Version="5.10.3" />
-        <PackageReference Include="GitHubActionsTestLogger" Version="1.2.0" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.10.0" />
-        <PackageReference Include="Serilog.Sinks.XUnit" Version="2.0.4" />
-        <PackageReference Include="Trakx.Utils.Testing" Version="0.2.0" />
+        <PackageReference Include="DotNetEnv" Version="2.3.0" />
+        <PackageReference Include="FluentAssertions" Version="6.5.1" />
+        <PackageReference Include="GitHubActionsTestLogger" Version="1.3.0">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+        <PackageReference Include="Serilog.Sinks.XUnit" Version="3.0.3" />
+        <PackageReference Include="Trakx.Utils.Testing" Version="0.3.1" />
         <PackageReference Include="xunit" Version="2.4.1" />
-        <PackageReference Include="Xunit.Extensions.AssemblyFixture" Version="2.2.0" />
+        <PackageReference Include="Xunit.Extensions.AssemblyFixture" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="coverlet.collector" Version="3.0.3">
+        <PackageReference Include="coverlet.collector" Version="3.1.2">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
@@ -33,6 +36,10 @@
       <ProjectReference Include="..\Trakx.Shrimpy.ApiClient\Trakx.Shrimpy.ApiClient.csproj" />
       <ProjectReference Include="..\Trakx.Shrimpy.Core\Trakx.Shrimpy.Core.csproj" />
       <ProjectReference Include="..\Trakx.Shrimpy.DeveloperApiClient\Trakx.Shrimpy.DeveloperApiClient.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Integration" />
     </ItemGroup>
 
 
