@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
@@ -9,7 +9,7 @@
     <Import Project="../Packable.Projects.props" />
 
     <ItemGroup>
-      <PackageReference Include="Trakx.Utils" Version="0.2.0" />
+      <PackageReference Include="Trakx.Utils" Version="0.3.1" />
         <PackageReference Include="Serilog" Version="2.10.0" />
     </ItemGroup>
 
