@@ -1,12 +1,10 @@
 ﻿
-namespace Trakx.Shrimpy.ApiClient
-{
-    internal partial class AccountsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
+namespace Trakx.Shrimpy.ApiClient;
 
+internal partial class AccountsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
 }
