@@ -3,41 +3,39 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Trakx.Shrimpy.Core;
 
-namespace Trakx.Shrimpy.ApiClient
+namespace Trakx.Shrimpy.ApiClient;
+
+public interface ITradabilityChecker
 {
-    public interface ITradabilityChecker
+    Task<IList<string>> GetTradableAsync(IEnumerable<string> symbols, CancellationToken cancellationToken = default);
+}
+
+public class TradabilityChecker : ITradabilityChecker
+{
+    private readonly IMarketDataClient _marketDataClient;
+
+    public TradabilityChecker(IMarketDataClient marketDataClient)
     {
-        Task<IList<string>> GetTradableAsync(IEnumerable<string> symbols, CancellationToken cancellationToken = default);
+        _marketDataClient = marketDataClient;
     }
 
-    public class TradabilityChecker : ITradabilityChecker
+    public async Task<IList<string>> GetTradableAsync(IEnumerable<string> symbols, CancellationToken cancellationToken = default)
     {
-        private readonly IMarketDataClient _marketDataClient;
-
-        public TradabilityChecker(IMarketDataClient marketDataClient)
+        var exchanges = ((IFavouriteExchangesClient)_marketDataClient).Top12ExchangeIds;
+        var untradableSymbols = new HashSet<string>(symbols);
+        foreach (var exchangeName in exchanges)
         {
-            _marketDataClient = marketDataClient;
-        }
-
-        public async Task<IList<string>> GetTradableAsync(IEnumerable<string> symbols, CancellationToken cancellationToken = default)
-        {
-            var exchanges = ((IFavouriteExchangesClient)_marketDataClient).Top12ExchangeIds;
-            var untradableSymbols = new HashSet<string>(symbols);
-            foreach (var exchangeName in exchanges)
+            if (Enum.TryParse(typeof(Exchange), exchangeName, true, out var exchange) && exchange != null)
             {
-                if (Enum.TryParse(typeof(Exchange), exchangeName, true, out var exchange) && exchange != null)
-                {
-                    var tickers = await _marketDataClient.GetTickerAsync((Exchange)exchange, cancellationToken);
-                    foreach (var ticker in tickers.Result)
-                        untradableSymbols.Remove(ticker.Symbol.ToLower());
+                var tickers = await _marketDataClient.GetTickerAsync((Exchange)exchange, cancellationToken);
+                foreach (var ticker in tickers.Result)
+                    untradableSymbols.Remove(ticker.Symbol.ToLower());
 
-                    if (untradableSymbols.Count == 0)
-                        break;
-                }
+                if (untradableSymbols.Count == 0)
+                    break;
             }
-            return symbols.Except(untradableSymbols).ToList();
         }
+        return symbols.Except(untradableSymbols).ToList();
     }
 }
